feat(classic): Connected the game to the hub

+++ server/internal/game/classic/bomb.go
@@ -4,7 +4,7 @@ import "github.com/N3moAhead/bomberman/server/pkg/types"
 
 type Bomb struct {
 	Pos  types.Vec2 `json:"pos"`
-	Fuse byte       `json:"fuse"` // Countdown each tick explodes at 0
+	Fuse int        `json:"fuse"` // Countdown each tick explodes at 0
 }
 
 func NewBomb(pos types.Vec2) *Bomb {
